Extract SnackbarHandler composable
